requirements.txt
absl-py==2.1.0
asttokens==3.0.0
audioread==3.0.1
backcall==0.2.0
ca-certificates==2024.12.14
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
comm==0.2.2
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
einops==0.8.0
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
ftfy==6.2.3
google-auth==2.34.0
google-auth-oauthlib==1.0.0
grpcio==1.66.1
huggingface-hub==0.25.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.12.2
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
krb5==1.21.3
lazy-loader==0.4
ld_impl_linux-64==2.38
libedit==3.1.20230828
libffi==3.4.4
libgcc==14.2.0
libgcc-ng==14.2.0
libgomp==14.2.0
librosa==0.10.2.post1
libsodium==1.0.20
libstdcxx==14.2.0
libstdcxx-ng==11.2.0
llvmlite==0.41.1
markdown==3.7
markupsafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
ml-collections==0.1.1
mpmath==1.3.0
msgpack==1.1.0
ncurses==6.4
nest-asyncio==1.6.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai-clip==1.0.1
opencv-python==4.10.0.84
openssl==3.4.0
packaging==24.1
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
pooch==1.8.2
prompt-toolkit==3.0.48
prompt_toolkit==3.0.48
protobuf==5.28.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pygments==2.18.0
pyparsing==3.1.4
python==3.8.19
python-dateutil==2.9.0.post0
python_abi==3.8
pytz==2024.2
pyyaml==6.0.2
pyzmq==25.1.2
readline==8.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.0
scipy==1.10.1
setuptools==72.1.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sqlite==3.45.3
stack_data==0.6.2
sympy==1.13.2
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1-2209072238
threadpoolctl==3.5.0
timm==1.0.9
tk==8.6.14
tokenizers==0.12.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.4.1
tqdm==4.65.0
traitlets==5.14.3
transformers==4.19.4
triton==2.1.0
typing-extensions==4.7.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.13
werkzeug==2.3.6
wheel==0.44.0
xz==5.4.6
zeromq==4.3.5
zipp==3.20.2
zlib==1.2.13
